Split GPU runtime and kernel families without changing dispatch or arithmetic
